Add support for resizing images on the fly
When assigning a photo to an email, a few common sizes are automatically
generated (and hardlinked), but when a different size is requested, the
mod_rewrite rules delegate to a Django resizer.

The Django resizer serves resized images but also saves them to disk so
that the second time an image is requested it is served by Apache
directly.