Add password reset functionality
All that's required to reset a password is an email address. This email
address must be confirmed in an account for the password reset to work.
(It's not possible to reset the password of an account which doesn't have
any confirmed email addresses. In these cases, a user can simply register
a new account and confirm the email addresses there.)

All validation errors will return the same generic error page. This is to
avoid leaking information about who is and isn't a user on the system.

The reset key is a hash of the username and the current password so that
we don't have to store this hash key in the database (and make it
expire). The same verification key will be sent if more than one password
reset is requested. However once the password is changed, that key will
automatically become invalid.