add_email: check first if it's already confirmed by a user
If an email address is already confirmed by another user, we still add it
to a user's list of unconfirmed email addresses, but we don't send an
email out with the verification key.

The reason for not returning an error message and refusing to add the
email to the account is to avoid exposing the fact that this email
address is already on the system.

I might revisit this in order to improve usability (a user forgetting
that he/she confirmed that email address in a different account).