Create a new directory for cropped files ("user" files)
This is in large part done to work around over-zealous HTTP caches which
ignore caching headers and can't tell that the image was resized/cropped.

One good side effect is that we can now have a cronjob which clears the
/uploaded/ directory for old files and so uploaded files which haven't
been cropped/resized will not hang around and will never be assigned to
email addresses.

Also moved the libravatar/avatar/image.py functions into other places to
get rid of that unnecessary directory.