Add email: put a limit on the number of unconfirmed emails
This is to put a bound on the data that can be stored per user account in
the database.

Also check whether this unconfirmed email has already been added or
whether it has been confirmed by another user.

If any of these checks fail, a non-descriptive error message is returned
to avoid leaking too much the fact that a given email address is in the
system.