Return a 404 when clients use www.libravatar.org/avatar/
Clients should be using cdn.libravatar.org/avatar/ otherwise the load
won't be spread across the mirrors. We could of course have a permanent
redirect to that server, however that still puts load on the main server
and clients won't be able to see that their code is broken when they are
doing their testing.