config: remove ETags on avatars since they are served by different hosts
This is a suggestion from Yahoo to increase cachability of files. Given
that ETags in Apache are by default specific to the host, it does not
play well with mirrors and CDNs and it's better to turn them off.

When ETags are missing, the usual "Last-Modified" headers will be used
instead.