Confirmed OpenIDs can now be used to login
In addition to adding a confirmed OpenID to one's list of OpenIDs that
can have a photo assigned to them, they are now also added to the list of
OpenIDs that can be used to login.

So a user can now have both a username and password and one or more
OpenIDs.