Change rsync path to ensure that top-level files get removed
rsync only delete from the destination the files that are missing on the
source host if the directory containing them is part of the rsync line.

"/var/lib/libravatar/avatar/*" doesn't include the avatar/ directory,
only the files contained within the directory.

"/var/lib/libravatar/avatar/" on the other hand includes the directory
and all of the files underneath that directory