dynamic resize: missing images should redirect to the nobody image
If an image has been removed from the master but the slave CDN still have
it, then there is a danger that the slave will redirect to the dynamic
resolver which will then redirect to the slave, etc.

This used to create infinite redirect loops for a short time after a
photo has been removed.