Turn ON DB-level autocommit and add @commit_on_success to view methods
Turning this ON means that transactions will be disabled by default
unless specified otherwise (with one of the Django ORM decorators). This
is good because it avoids unnecessarily wrapping read methods with
transactions.

  http://djangodocs.com/1.3/s-api-reference/databases/#s-autocommit-mode

Most view methods that write to the DB on the other hand must have a
commit_on_success or commit_manually decorator.

  http://djangodocs.com/1.3/s-using-django/s-module-django.db/s-module-django.db.transaction/s-controlling-transaction-management-in-views/#django.db.transaction.commit_on_success