Automatic creation of accounts through BrowserID (bug #776016)
If a user tries to login using a BrowserID that doesn't match a confirmed
email on the site, then a new account is automatically created. The
process is pretty much the same as for the OpenID auth backend.

Also include email addresses (in case a BrowserID could be used) when a
user attempts to reset the password in a password-less account.