Convert the add_email page to use the Observer API
Because we are using request() for something else than what it was
intended, we need to silently swap our browserid_user variable to
match the new state that BrowserID keeps.

Specifically, calling request() changes the currently logged in
user as far as BrowserID is concerned and the onlogin event will
get called so we need to override it to make sure that the user
isn't logged out and back again (which would create a new account
with the email that was supposed to be added to the existing
account).