When adding an email that's taken, clear browserid_user
Because we are abusing the request() API to add alternate emails
to one's account, we need to make sure that BrowserID doesn't get
into a situation where it thinks that a user is logged whereas
that loging was aborted by Libravatar.

Therefore, when an add_email operation fail because another account
already has that email, Libravatar needs to silently log that user
out of BrowserID. Furthermore, it needs to clear its browserid_user
variable which is now out of sync with BrowserID's state.