Call navigator.id.logout() when removing browserid_user email
To ensure that the BrowserID internal session matches reality, we
need to silently log the user out of BrowserID when the email that
is being removed matches the one that BrowserID thinks was used to
login in this session.

Doing so means that Libravatar switches to a non-BrowserID session.

Took this opportunity to simplify the common "silent logout" code
with the "delete account" operation.