Generate new pot file for Persona strings
Also accept fuzzy strings in the French translation since all that
changed is the name of the product (from BrowserID to Persona).

It is useful to note that the name BrowserID is still present in
error messages since in that case, it mostly refers to the protocol
and not to the service.